Quest repeat fixes (#552)
* Iterate through all teams, and not just named teams in certain circumstances

* Reset instant-repeatable quests with other repeatables to fix quest data stuck in a completed state

* Change QuestTicker to use game-time from the world to get a world-persistent and synced time unit

* Fix deserialization of RepeatInfo

* Fix cooldown time text and cleanup repeat time text code slightly

* Make tick event listening on forge platform match fabric platform by only listening to tick end

+++ common/src/main/java/hardcorequesting/common/io/adapter/QuestDataAdapter.java
@@ -68,7 +68,7 @@ public class QuestDataAdapter {
                         data.available = in.nextBoolean();
                         break;
                     case TIME:
-                        data.time = in.nextInt();
+                        data.time = in.nextLong();
                         break;
                     case TASKS_SIZE:
                         data.tasks = new QuestDataTask[in.nextInt()];
